Moves DrawerIconButton into screen folder since it's specific to that screen.
